# openapi3.yaml
openapi: 3.0.1
info:
  title: Layer-Spec
  description: >-
    exposes CRUD (Create, Read, Update and Delete) operations into/from
    layer-spec database.
  version: 1.1.2
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
paths:
  /tilesCount/{layerId}/{target}:
    get:
      operationId: getTilesCount
      tags:
        - tiles
      summary: gets the tiles count by layer id
      parameters:
        - name: layerId
          in: path
          required: true
          schema:
            type: string
        - name: target
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getTilesCount'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    put:
      operationId: updateTilesCount
      tags:
        - tiles
      summary: insert/update the tiles count by layer id
      parameters:
        - name: layerId
          in: path
          required: true
          schema:
            type: string
        - name: target
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/upsertTilesCountReq'
      responses:
        '201':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/upsertTilesCountRes'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
components:
  schemas:
    error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    getTilesCount:
      type: object
      required:
        - tilesCount
      properties:
        tilesCount:
          type: number
    upsertTilesCountRes:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    upsertTilesCountReq:
      type: object
      required:
        - tilesBatchCount
      properties:
        tilesBatchCount:
          type: integer